organize project folder structure
