+++ Frontend/src/assets/img/spotify_menu_logo.png
